Removed `ccLogBase.py`: deprecated core logging utility as its functionality is now streamlined and handled by updated modular components in `WrenchCL`.
